moved buttons to settings activity
