+++ Breathe/app/src/main/res/layout/settings.xml
@@ -9,14 +9,47 @@
         android:id="@+id/toolbar"
         layout="@layout/toolbar"/>
 
-    <Button
-        android:id="@+id/button"
-        android:layout_width="wrap_content"
+    <LinearLayout
+        android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:text="Button"
-        app:layout_constraintBottom_toBottomOf="parent"
+        android:layout_marginStart="@dimen/default_margin"
+        android:layout_marginTop="@dimen/default_margin"
+        android:layout_marginEnd="@dimen/default_margin"
+        android:orientation="vertical"
         app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="1.0"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        android:onClick="exportAllIUEs"/>
+        app:layout_constraintTop_toBottomOf="@+id/toolbar">
+
+        <Button
+            android:id="@+id/scan_wearable_button"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:onClick="onScanButtonClick"
+            android:text="@string/connect_to_wearable" />
+
+        <Button
+            android:id="@+id/export_all_iue_button"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:onClick="exportAllIUEs"
+            android:text="@string/export_data_as_csv" />
+
+        <!-- delete later. Leave hardcoded string-->
+        <Button
+            android:id="@+id/simulate_iue_button"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:onClick="onSimulateIUEButtonClick"
+            android:text="Simulate IUE" />
+
+        <Button
+            android:id="@+id/test_wearable_button"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:onClick="testWearableData"
+            android:text="Test wearable" />
+    </LinearLayout>
+
+
 </androidx.constraintlayout.widget.ConstraintLayout>
