+++ Breathe/app/src/main/res/values/strings.xml
@@ -9,4 +9,6 @@
     <string name="placeholder_text">-</string>
     <string name="diary_label">Diary</string>
     <string name="settings_label">Settings</string>
+    <string name="connect_to_wearable">Connect to wearable</string>
+    <string name="export_data_as_csv">Export data as csv</string>
 </resources>
